+++ libs/components/tabs/testing/src/modules/tabs/tab-button-harness.ts
@@ -46,6 +46,9 @@ export class SkyTabButtonHarness extends SkyComponentHarness {
    * Clicks the remove tab button if it is visible.
    */
   public async clickRemoveButton(): Promise<void> {
+    if (await this.isInWizardTabset()) {
+      throw new Error('Cannot use remove tab button in a wizard tabset.');
+    }
     const button = await this.locatorForOptional('.sky-btn-tab-close')();
     if (!button) {
       throw new Error('Unable to find remove tab button.');
@@ -57,6 +60,11 @@ export class SkyTabButtonHarness extends SkyComponentHarness {
    * Gets the permalink that the page routes to when the tab is clicked.
    */
   public async getPermalink(): Promise<string | null> {
+    if (await this.isInWizardTabset()) {
+      throw new Error(
+        'Cannot get permalink for tab button in a wizard tabset.',
+      );
+    }
     return await (await this.#getTabButton()).getAttribute('href');
   }
 
@@ -91,6 +99,13 @@ export class SkyTabButtonHarness extends SkyComponentHarness {
    */
   public async isDisabled(): Promise<boolean> {
     return await (await this.#getTabButton()).hasClass('sky-btn-tab-disabled');
+  }
+
+  /**
+   * Whether the tab button is in a Wizard tabset.
+   */
+  public async isInWizardTabset(): Promise<boolean> {
+    return await (await this.#getTabButton()).hasClass('sky-btn-tab-wizard');
   }
 
   /**